+++ Tests/Caravela.Framework.Tests.PublicPipeline/Tests/Aspects/Introductions/Properties/ExistingConflictNew.t.cs
@@ -23,105 +23,105 @@ return 27;    }
 public new global::System.Int32 BaseClassProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.BaseClassProperty;
+        return base.BaseClassProperty;
     }
 }
 
 public static new global::System.Int32 BaseClassProperty_Static
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)global::Caravela.Framework.IntegrationTests.Aspects.Introductions.Properties.ExistingConflictNew.BaseClass.BaseClassProperty_Static;
+        return global::Caravela.Framework.IntegrationTests.Aspects.Introductions.Properties.ExistingConflictNew.BaseClass.BaseClassProperty_Static;
     }
 }
 
 public new global::System.Int32 HiddenBaseClassProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.HiddenBaseClassProperty;
+        return base.HiddenBaseClassProperty;
     }
 }
 
 public static new global::System.Int32 HiddenBaseClassProperty_Static
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)global::Caravela.Framework.IntegrationTests.Aspects.Introductions.Properties.ExistingConflictNew.DerivedClass.HiddenBaseClassProperty_Static;
+        return global::Caravela.Framework.IntegrationTests.Aspects.Introductions.Properties.ExistingConflictNew.DerivedClass.HiddenBaseClassProperty_Static;
     }
 }
 
 public new global::System.Int32 HiddenBaseClassVirtualProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.HiddenBaseClassVirtualProperty;
+        return base.HiddenBaseClassVirtualProperty;
     }
 }
 
 public new global::System.Int32 HiddenVirtualBaseClassVirtualProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.HiddenVirtualBaseClassVirtualProperty;
+        return base.HiddenVirtualBaseClassVirtualProperty;
     }
 }
 
 public new global::System.Int32 BaseClassVirtualProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.BaseClassVirtualProperty;
+        return base.BaseClassVirtualProperty;
     }
 }
 
 public new global::System.Int32 BaseClassVirtualSealedProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.BaseClassVirtualSealedProperty;
+        return base.BaseClassVirtualSealedProperty;
     }
 }
 
 public new global::System.Int32 BaseClassVirtualOverridenProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.BaseClassVirtualOverridenProperty;
+        return base.BaseClassVirtualOverridenProperty;
     }
 }
 
 public new global::System.Int32 BaseClassAbstractProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.BaseClassAbstractProperty;
+        return base.BaseClassAbstractProperty;
     }
 }
 
 public new global::System.Int32 BaseClassAbstractSealedProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.BaseClassAbstractSealedProperty;
+        return base.BaseClassAbstractSealedProperty;
     }
 }
 
 public new global::System.Int32 DerivedClassProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.DerivedClassProperty;
+        return base.DerivedClassProperty;
     }
 }
 
 public static new global::System.Int32 DerivedClassProperty_Static
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)global::Caravela.Framework.IntegrationTests.Aspects.Introductions.Properties.ExistingConflictNew.DerivedClass.DerivedClassProperty_Static;
+        return global::Caravela.Framework.IntegrationTests.Aspects.Introductions.Properties.ExistingConflictNew.DerivedClass.DerivedClassProperty_Static;
     }
 }
 
 public new global::System.Int32 DerivedClassVirtualProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.DerivedClassVirtualProperty;
+        return base.DerivedClassVirtualProperty;
     }
 }
 
 public new global::System.Int32 DerivedClassVirtualSealedProperty
 {get    {
         global::System.Console.WriteLine("This is introduced property.");
-        return (int)base.DerivedClassVirtualSealedProperty;
+        return base.DerivedClassVirtualSealedProperty;
     }
 }
